pyproject.toml
[tool.poetry]
name = "rag-arxiv"
version = "0.1.0"
description = ""
authors = [
    "Emerson Lopes <eplopes@inf.ufpel.edu.br>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "==2.9.10"
python-decouple = "^3.8"
pypdf = "^5.1.0"
google-generativeai = "^0.8.3"
pgvector = "^0.3.6"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
